pyproject.toml
[project]
name = "parsedf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.0.0",
    "pymupdf>=1.25.1",
    "pyqt5>=5.15.11",
]

[dependency-groups]
dev = [
    "pyright>=1.1.391",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]
